requirements.txt
cryptography>=44.0.0
paramiko>=3.5.0
PyYAML>=6.0.2
scp>=0.15.0